DPTB-79 refactor: extract NotificationSenderService from NotificationServiceImpl
