+++ src/main/kotlin/ru/illine/drinking/ponies/service/notification/NotificationSenderService.kt
@@ -1,0 +1,11 @@
+package ru.illine.drinking.ponies.service.notification
+
+import ru.illine.drinking.ponies.model.dto.internal.NotificationSettingDto
+
+interface NotificationSenderService {
+
+    fun sendNotifications(notifications: Collection<NotificationSettingDto>)
+
+    fun suspendNotifications(notifications: Collection<NotificationSettingDto>)
+
+}
