+++ src/main/kotlin/ru/illine/drinking/ponies/service/notification/NotificationService.kt
@@ -1,7 +1,6 @@
 package ru.illine.drinking.ponies.service.notification
 
 import org.telegram.telegrambots.abilitybots.api.objects.MessageContext
-import ru.illine.drinking.ponies.model.dto.internal.NotificationSettingDto
 
 interface NotificationService {
 
@@ -15,7 +14,4 @@ interface NotificationService {
 
     fun settings(messageContext: MessageContext)
 
-    fun sendNotifications(notifications: Collection<NotificationSettingDto>)
-
-    fun suspendNotifications(notifications: Collection<NotificationSettingDto>)
 }
